+++ adapter/ph/src/agent_output_worker.rs
@@ -4,8 +4,9 @@ use crate::counters::*;
 use crate::fastpath;
 use crate::net_defs;
 use crate::packet::Packet;
+use crate::sys::TunPi;
 use crate::sys::ZprTun;
-use crate::zprtun::tun_pi;
+use crate::zprtun;
 use std::sync::Arc;
 
 #[derive(Copy, Clone)]
@@ -15,7 +16,7 @@ pub struct Config {
     pub batch_size: usize,
 }
 
-fn is_ip(pi: tun_pi::TunPi) -> bool {
+fn is_ip(pi: TunPi) -> bool {
     pi.proto == net_defs::ethertype::IP || pi.proto == net_defs::ethertype::IPV6
 }
 
@@ -33,15 +34,23 @@ pub async fn launch(config: Config, asm: Arc<Assembly>, tun: Arc<ZprTun>) {
         for mut buf in bufs.drain(..) {
             let pkt = loop {
                 let mut pkt = Packet::new(buf, config::DEFAULT_MESSAGE_HEADROOM);
-
                 tun.recv_buf(&mut pkt).await.unwrap();
-                let pi = tun_pi::read_pi(&mut pkt);
-                if pi.strip || !is_ip(pi) {
-                    // packet was too large or non-IP; drop
-                    asm.counters[CounterType::OutPacksDrop].increment();
-                    // reuse `buf`
-                    buf = pkt.destroy();
-                    continue;
+                if zprtun::TUN_HAS_PI {
+                    let pi = TunPi::read_pi(&mut pkt);
+                    if pi.strip || !is_ip(pi) {
+                        // packet was too large or non-IP; drop
+                        asm.counters[CounterType::OutPacksDrop].increment();
+                        // reuse `buf`
+                        buf = pkt.destroy();
+                        continue;
+                    }
+                } else {
+                    // No packet info, permit IP and IPv6 only (for now?)
+                    if pkt.body()[0] >> 4 != 4 && pkt.body()[0] >> 4 != 6 {
+                        asm.counters[CounterType::OutPacksDrop].increment();
+                        buf = pkt.destroy();
+                        continue;
+                    }
                 }
 
                 break pkt;
