+++ adapter/ph/src/sys/macos.rs
@@ -1,2 +1,4 @@
 pub mod zprtun;
 pub use zprtun::ZprTun;
+pub mod tun_pi;
+pub use tun_pi::TunPi;
